Lucubration Builder/requirements.txt
PyYAML==6.0.1
pyinstaller==6.9.0
